I added a couple pages to replace the booking feature
Requests page will show users requests to walkers. When a user clicks the request button on a walker from the walkers list the request will be added to the database and will show up on that walkers request page. I also added the navigation bar to the rest of the pages.